Nettoyage des methodes, correction d'erreur de type dans Client

+++ src/main/java/al420445/models/Users/Client.java
@@ -26,7 +26,7 @@ public class Client {
 
     @Id
     @GeneratedValue
-    private int clientID;
+    private long clientID;
 
     private String clientName;
 
@@ -47,7 +47,7 @@ public class Client {
 
 
 
-    public int getClientID() {
+    public long getClientID() {
         return clientID;
     }
 
@@ -101,7 +101,6 @@ public class Client {
     private boolean checkHasDebts(){
         return !this.dettes.isEmpty();
     }
-
 
 
 }
